Move telemetry upload code to core, likewise with tests
