+++ python_modules/dagit/dagit/cli.py
@@ -1,8 +1,6 @@
 import logging
 import os
 import sys
-import threading
-from contextlib import contextmanager
 from typing import Optional
 
 import click
@@ -14,6 +12,7 @@ from dagster.cli.workspace import (
 )
 from dagster.cli.workspace.cli_target import WORKSPACE_TARGET_WARNING
 from dagster.core.telemetry import START_DAGIT_WEBSERVER, log_action
+from dagster.core.telemetry_upload import uploading_logging_thread
 from dagster.core.workspace import WorkspaceProcessContext
 from dagster.utils import DEFAULT_WORKSPACE_YAML_FILENAME
 from dagster.utils.log import configure_loggers
@@ -21,7 +20,6 @@ from gevent import pywsgi
 from geventwebsocket.handler import WebSocketHandler
 
 from .app import create_app_from_workspace_process_context
-from .telemetry import upload_logs
 from .version import __version__
 
 
@@ -167,21 +165,6 @@ def host_dagit_ui_with_workspace_process_context(
     app = create_app_from_workspace_process_context(workspace_process_context, path_prefix)
 
     start_server(workspace_process_context.instance, host, port, path_prefix, app, port_lookup)
-
-
-@contextmanager
-def uploading_logging_thread():
-
-    stop_event = threading.Event()
-    logging_thread = threading.Thread(
-        target=upload_logs, args=([stop_event]), name="telemetry-upload"
-    )
-    try:
-        logging_thread.start()
-        yield
-    finally:
-        stop_event.set()
-        logging_thread.join()
 
 
 def start_server(instance, host, port, path_prefix, app, port_lookup, port_lookup_attempts=0):
